Illustrate hashchange
hashchange does not fire when the page first loads. It only fires when
the hash changes from what it was on the last load. You can test this by
loading the page at hashchange/index.html with a URL ending like
hashchange/index.html#abc and observing that 'hash changed to abc' is
not logged. Then you can change the end of the URL to
hashchange/index.html#xyz, and observe that 'hash changed to xyz' is
logged.